Move deptrace setup into a method in its own file
- Add ability deptrace setup to use localVersion from deps.state instead of
  expectedVersion.
