+++ lib/status.js
@@ -1,7 +1,7 @@
 const deps = require('./util/deps');
 
 module.exports = function(dir) {
-  return this.graph(dir).then(deps.visualize).then(function(lines) {
+  return this.graphTarget(dir).then(deps.visualize).then(function(lines) {
     lines.forEach(function(line) {
       this.emit('write', line);
     }.bind(this));
